Extract CompletedStreamedResponse from three duplicate wrapper classes
TemporalStreamedResponse, PrefectStreamedResponse, and DBOSStreamedResponse
were identical implementations. Consolidate into a single shared class in
models/wrapper.py. Mechanical: no behavior change.
